+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   - If this header is still required by your implementation, you need to implement a custom SmtpMailer. 
 - Added string error code to the Subscribe APIs to differentiate between different 404 scenarios. remp/web#2263 
 - Fix Mailer segment provider users acquiring. Provided segment code needs to be processed before fetching users from database. remp/mnt#114
+- Fix New template generator form - broken sorting value `after`. If selected, select box was not shown. remp/helpdesk#2073
 
 ## Archive